Gauntlet 1.4
Added a RedVignette widget to make the screen flash red when the player takes damage. Created an animation for the widget that sets opacity of the image from 0 to 1 to 0. Then used a Remove from Parent node to delete each instantiation of the widget to improve performance. Third person player blueprint plays this animation when the player takes damage.